Add delete_from_memcache option to SimpledbResource#delete_all.
Change sdb_backup to not delete from memcache when deleting old rows.

git-svn-id: https://tapjoy.unfuddle.com/svn/tapjoy_tapjoyads/trunk@3454 6df48309-bd3c-4361-8df8-ec25136b6021

+++ tapjoyads/app/models/simpledb_resource.rb
@@ -308,8 +308,8 @@ class SimpledbResource
   
   ##
   # Deletes this entire row immediately (no need to call save after calling this).
-  def delete_all
-    delete_from_cache(get_memcache_key)
+  def delete_all(delete_from_memcache = true)
+    delete_from_cache(get_memcache_key) if delete_from_cache
     @@sdb.delete_attributes(@this_domain_name, key)
   end
